Suggest additional link
I don't know if you're taking requests, but this is a detailed and accessible article (there's an accompanying presentation as well, same content) on how namespaces and containers work.  If so, here you go.  (Disclosure: Yes, I wrote it, but it's gotten very good traction at conferences in the past few years.)